# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TOPIDEALSPACE API",
        "description": "API REST de la marketplace TOPIDEALSPACE. Authentification via Laravel Sanctum (Bearer Token). Connectez-vous via **POST /auth/login**, copiez le token retourné, puis cliquez sur **Authorize** en haut à droite.",
        "contact": {
            "email": "contact@topidealspace.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local (dev)"
        }
    ],
    "paths": {
...
        "/admin/profile": {
            "get": {
                "tags": [
                    "Admin Profil"
                ],
                "summary": "Profil de l'admin connecté",
                "description": "Récupère les informations du profil de l'administrateur authentifié",
                "operationId": "adminGetProfile",
                "responses": {
                    "200": {
                        "description": "Informations du profil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_nom": {
                                                    "type": "string"
                                                },
                                                "user_prenom": {
                                                    "type": "string"
                                                },
                                                "user_email": {
                                                    "type": "string"
                                                },
                                                "user_login": {
                                                    "type": "string"
                                                },
                                                "user_tel": {
                                                    "type": "string"
                                                },
                                                "user_image": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "user_email_status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Profil"
                ],
                "summary": "Mettre à jour le profil",
                "description": "Met à jour les informations de profil de l'administrateur connecté",
                "operationId": "adminUpdateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_nom",
                                    "user_prenom"
                                ],
                                "properties": {
                                    "user_nom": {
                                        "description": "Nom",
                                        "type": "string",
                                        "maxLength": 150
                                    },
                                    "user_prenom": {
                                        "description": "Prénom",
                                        "type": "string",
                                        "maxLength": 30
                                    },
                                    "user_tel": {
                                        "description": "Téléphone",
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/profile/image": {
            "post": {
                "tags": [
                    "Admin Profil"
                ],
                "summary": "Mettre à jour la photo de profil",
                "description": "Upload et met à jour la photo de profil de l'administrateur",
                "operationId": "adminUpdateProfileImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Image de profil (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "image_url": {
                                                    "type": "string"
                                                },
                                                "image_width": {
                                                    "type": "integer"
                                                },
                                                "image_height": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Profil"
                ],
                "summary": "Supprimer la photo de profil",
                "description": "Supprime la photo de profil et remet l'avatar par défaut",
                "operationId": "adminDeleteProfileImage",
                "responses": {
                    "200": {
                        "description": "Photo supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/profile/password": {
            "put": {
                "tags": [
                    "Admin Profil"
                ],
                "summary": "Changer le mot de passe",
                "description": "Change le mot de passe de l'administrateur. Révoque tous les tokens existants.",
                "operationId": "adminUpdatePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "description": "Mot de passe actuel",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "Nouveau mot de passe",
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "new_password_confirmation": {
                                        "description": "Confirmation du nouveau mot de passe",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié. Reconnexion requise."
                    },
                    "422": {
                        "description": "Mot de passe actuel incorrect ou validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum Bearer Token. Entrez le token obtenu lors de la connexion.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}